Attempt a GWT-friendly encrypted FileHandle.
It seems very resistant to the idea of working... If any filtering is done on an encrypted text file, it still probably is treated as text, so CR will change to LF and the encryption will break. The files need to be served as binary, which doesn't play nicely with GWT's Preloader (especially for images).